Unify table types and rename as simply table
